// app/dashboard/requisitions/types/requisition-type.enum.ts
export enum RequisitionType {
    ADJUSTMENT = "ADJUSTMENT",
    PURCHASE_RECEIPT = "PURCHASE_RECEIPT",
    RENT = "RENT",
    TRANSFER = "TRANSFER",
    RETURN = "RETURN",
    CONSUMPTION = "CONSUMPTION",
    SALE = "SALE",
    INTERNAL_TRANSFER = "INTERNAL_TRANSFER",
    MAINTENANCE = "MAINTENANCE",
    OUT_OF_SERVICE = "OUT_OF_SERVICE",
}